Added a bunch of items! Crafting items, consumable items (including steroids), and even made the display for said items mimic the game's!
Also moved a bunch of repetitive tables and such into a function that generates them onload. Ideally there'd be a server that did this instead of JSing it at runtime, but it's not a huge deal for this project.

Also apparently the folder structure is still screwed up but it doesn't effect the site itself so it's fine enough I guess.